remove some no-longer-needed modules
