+++ nobodywho/core/src/lib.rs
@@ -1,9 +1,9 @@
 // pub mod chat;
+pub mod chat;
 pub mod chat_state;
 pub mod embed;
 pub mod llm;
 // pub mod llmactor;
-pub mod chatworker;
 pub mod sampler_config;
 
 #[cfg(test)]
